# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShopademiaStore API
  version: 1.0.0
  description: >-
    Inventory of the routes currently wired in backend/routes/v1 and the
    product-media static handler. Request and response field schemas are marked
    unknown where the implementation contract has not been fully specified here.
    CORS preflight OPTIONS handling is middleware-generated rather than an
    explicit router declaration and is not separately modeled as an operation.
servers:
  - url: /
    description: API origin; versioned API paths begin with /api/v1.
...
paths:
...
  /api/v1/admin/auth/login:
    post:
      tags: [Admin authentication]
      operationId: adminLogin
      summary: Authenticate an administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request field schema is unknown here; see backend/validators/admin-auth.validators.js.
      responses:
        "200":
          description: Authentication success envelope.
        "401":
          description: Invalid credentials.
        "429":
          description: Login rate limit exceeded.
  /api/v1/admin/auth/me:
    get:
      tags: [Admin authentication]
      operationId: getAdminIdentity
      security:
        - AdminBearer: []
      responses:
        "200":
          description: Authenticated administrator success envelope.
        "401":
          description: Authentication required.
        "403":
          description: Role not permitted.
...
  /api/v1/auth/login:
    post:
      tags: [Customer authentication]
      operationId: loginCustomer
      security:
        - StoreKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { description: "Login fields are defined by backend/validators/customer-auth.validators.js; field schema is not duplicated here." }
      responses:
        "200":
          description: Customer authentication success envelope.
        "401":
          description: Invalid credentials.
        "429":
          description: Login rate limit exceeded.
...
components:
  securitySchemes:
    StoreKey:
      type: apiKey
      in: header
      name: X-Store-Key
      description: Public store identifier; not an authorization credential for administrative actions.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Administrator access token.
    CustomerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Store-bound customer access token.
...